refactor(channels): split claim-write retries out of ingress-drain
Recently landed #127849 and #127950 grew ingress-drain.ts past the
max-lines cap between each other's CI runs, so every merge ref now fails
check-lint-core-4. Move the bounded claim-write retries and the
complete/release/dead-letter wrappers into ingress-claim-writes.ts —
byte-identical functions, unchanged public surface and behavior — taking
the file from 703 to 625 counted lines with headroom, no suppression.
